+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
 - Enh #399: Remove deprecated widget `DataSaved`
 - Enh #404: Copy templates
 - Enh #406: Twig extension to process markdown text
+- Enh #410: Allow to export/import whole container
 
 1.10.16 (Unreleased)
 ---------------------------